backend/requirements.txt
django
djangorestframework
python-decouple
celery
redis
openai>=1.3.5
pillow
